requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bleach==3.3.1
Django==4.0.3
django-ckeditor==6.2.0
django-classy-tags==3.0.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-nyt==1.2.2
django-sekizai==3.0.1
Markdown==3.2.2
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
pyparsing==3.0.7
python-dotenv==0.19.2
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
webencodings==0.5.1
